+++ NEWS.md
@@ -1,6 +1,7 @@
 # Changes since last release
 
 * add argument "deuOnly" to substSpecialChars()
+* add function clipMatrix()
 
 # [kwb.utils 0.11.0](https://github.com/KWB-R/kwb.utils/releases/tag/v0.11.0) <small>2022-01-19</small>